feat: Added further flag usage to all scripts

+++ stop.sh
@@ -1,25 +1,29 @@
 #!/bin/bash
 
-function prompt() {
-  clear
-  echo "Do you want to: "
-  echo "1. Only stop the container"
-  echo "2. Stop and remove the container"
-  echo "3. Exit"
-  read command
+help() {
+  echo "Usage: $0 [OPTIONS]"
+  echo "Options:"
+  echo "-h, --help       Display help message"
+  echo "-rm, --remove     Stop and remove docker compose containers"
 }
 
-prompt
 
-while [[ $command -ne 3 ]]; do
-  case $command in
-    1)
-      docker compose stop
-      ;;
-    2)
-      docker compose down
-      ;;
-  esac
-  prompt
-done
+  if [ $# -gt 0 ]; then
+    case $1 in
+      -h | --help)
+        help
+        exit 0
+        ;;
+      -rm | --remove)
+        docker compose down
+        ;;
+      *)
+        echo "Invalid argument"
+        help
+        exit 0
+        ;;
+    esac
+  else :
+    docker compose stop
+  fi
